:gift: BONUS: Display QA term instead of id in search results and facets
This commit adds helper methods to department so that the value will display the terms instead of the id, in the catalog search index and facets. For example, "history" instead of "hst". You could write a spec for these methods for reassurance, and know that this is not the only way to accomplish this. With customizations the dev will often have to come up with solutions of their own.

:gift: BONUS: Display QA term instead of id in search results and facets

This commit adds helper methods to department so that the value will display the terms instead of the id, in the catalog search index and facets. For example, "history" instead of "hst". You could write a spec for these methods for reassurance, and know that this is not the only way to accomplish this. With customizations the dev will often have to come up with solutions of their own.