🎁 Maintain Input order for Image and GenericWork's creator
Prior to this commit, there was no guarantee that the creator values
fetched from Fedora would come back in the same order (as RDF is by
nature unnorder).

With this commit, we're introducing the OrderAlready gem to sort the
creator field based on the input order.  That is to say when we specify
the "first" creator (e.g. first author) that will always return as the
first creator.

Note, what's happening under the hood is that we attach an ordering
prefix to the value stored in Fedora.  So, if you're referencing Fedora
outside of this Hyrax-based application, you might see things like:

Creator "0~Jeremy Friesen"